requirements.txt
cfgv==3.4.0
distlib==0.3.9
filelock==3.18.0
identify==2.6.12
iniconfig==2.1.0
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
tabulate==0.9.0
virtualenv==20.31.2
